Server: Move security config class to dedicated packages and update them to package private and final when possible.
