--- server/src/main/java/com/gms/component/security/authentication/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains classes to handle the session state.
- */
-package com.gms.component.security.authentication;
