+++ server/src/main/java/com/gms/config/locale/package-info.java
@@ -1,7 +1,7 @@
 /**
- * This package contains classes for handling i18n and l10n.
+ * This package contains configuration classes for handling i18n and l10n.
  *
  * @author Asiel Leal Celdeiro | lealceldeiro@gmail.com
- * @version 0.1
+ * @version 0.2
  */
 package com.gms.config.locale;
