+++ server/src/main/java/com/gms/config/package-info.java
@@ -2,6 +2,6 @@
  * This package contains classes regarding springboot configuration.
  *
  * @author Asiel Leal Celdeiro | lealceldeiro@gmail.com
- * @version 0.1
+ * @version 0.2
  */
 package com.gms.config;
