+++ server/src/main/java/com/gms/config/security/authorization/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains classes in charge of handling the authorization flow.
+ *
+ * @author Asiel Leal Celdeiro | lealceldeiro@gmail.com
+ * @version 0.1
+ */
+package com.gms.config.security.authorization;
